rules: use bulk_create to fasten insert
Importing a huge ruleset is slow because of the number of queries
done. We try to use bulk_create in that patch to reduce that
number of queries.

Using a Django object to store Reference linked to a Rule was
preventing the bulk creation from working because ManyToMany
relation are not supported. This patch suppresses this object
from Model and uses a direct parsing of Rule content when
references have to be displayed. As references are only displayed
in the rule detail, there is no penality in term of performance.

With this patch updating ETOpen Ruleset on sqlite3 goes from
45 minutes to less than 25 secondes.