rules: add two management commands
This patch adds two management commands: 'addsource' and
'defaultruleset'. The idea is to be able to create from
command line objects in the database to have a ready to
use scirius (for example in the case of SELKS).

'addsource' can be used to add and update a source. Its syntax is as
follow:
 python manage.py addsource "ETOpen Ruleset" https://rules.emergingthreats.net/open/suricata-2.0.1/emerging.rules.tar.gz http sigs

'defaultruleset' will create a ruleset using all defined sources and
taking all categories available in the sources. The syntax is as
follow:

 python manage.py defaultruleset "ETOpen based ruleset"