rules: improve timeline code
JS code has been moved in scirius.js file. The system now support
filtering by host. It also displays cumulative line if multiple
hosts are asked.

On low traffic box, the bar char was inconsistent and the plot is
now done with a line. This way we have a correct time axis.