rules: fix missing 0 issue in graph
Elasticsearch is not returning an entry when there is no data for
the interval. This confused nvd3 which needs regular tick. This
patch updates the JS code to add data the missing values.

This allows to switch to the bar representation which is more
conform to the reality of the represented data.