scirius: add reverse proxy capability
Kibana and Elasticsearch are not by default protected by
authentication. To avoid a new layer of authentification, scririus
is now able to reverse proxy connections to Kibana and
Elasticsearch. This makes a cheap SSO solution. If you want to
implement this you will need to setup your Elasticsearch and Kibana
to have them not accessible from outside. Scirius will then be able
to be used as the only gateway.