rules: improve source addition
Usually a user will want to add all categories of a source to some
ruleset when it adds a source. This patch does so by giving to the
user the possibility to select in which rulesets the source must
be fully added.

As a side effect, a page is introduced to display a progress bar
and describe the operation done during initialisation.