rules: user can choose to continue
During a source addition, we give the user the choice to destroy
the source or to continue even if there is errors during the test.
Last choice is being given because we trust user but also because
error could come to the fact a file from another source is used.