rules: optimize activity checking
Query generated by django in SQL is really slow. A specific setup
was seeing 400ms time for request execution causing a delay of
around 1.5s to generate the page (because 3 rulesets were used).
By using a direct test we simplify the SQL queries and get back
to a loading time of the rule page below 100ms.