rules: fix ES index list building
In the case where from_date is in the previous index, then from_date
added to interval can be greated than now. Result was last existing
Elasticsearch index not to be used in some cases.

This was causing at least trend to be wrongly computed in this case.
This may also fix some other problems such as no data in some graph
when the inteval was too big.