rules: cache category when building ruleset
Performance of ruleset generation was impacted by the IPS work with
a run in 10 sec instead of 1 sec. All execution time was spend in
the is_transformed() function. This was due to the fact, that to
test if the category is in a transformation list we had a database
access for each rule.

This patch uses select_related() Django function to fetch the
category when building the set of rules. This way we hit the query
cache when doing the category check. This was enough to get back to
a 1 sec duration run.