rules: fix rule update strategy
In the case where a ruleset has no flowbits, the updated rules was
containing the whole set of rules. This was linked with the fact
if there is no flowbit then we try to initialize them at the source
level. And in that case it was triggering a global update of rules.

This patch change the way we update rule by only considering an
update is needed (due to flowbits) if we are initializing flowbits
and we have a flowbit in the rule.