rules: conditional rule transform
Some rules can't be transformed. For instance a rule with flowbits
at noalert can not be set to drop as it will cause the set of
depending rules to be broken. This is the same for filestore if
a rule is not http and smtp.

This patch updates the code to have a dynamic list of transformations
per rule and disable the transformation if no transformation is
available.