rules: improve tar file handling
Tar file were currently force t ohave a structure with a 'rules'
directory on top of the archive. This patch updates the structure
to be able to have one level 1 directory followed by a rules
directory.

As a consequence of this change, the patch is also updating the
way categories are build by looking into the filesystem instead
of looking into the tar file (once again).