rules: fix ES index expansion
When the index list is too long we get back to a star index but
by using `logstash-*` we end up looking in all indexes. So this
patch appends a 2 seen as start of years. Result is that ES is
only looking in the stats index and we avoid serious performances
problems.