hunt: graph rerender fixed
Originally the Chart.js component was using c3js-react wrapper component for managing the graph.
There are some known bugs for this wrapper and one of them is actually that the graph is being incorrectly updated
Instead of using this wrapper I switched to original c3 module where these bugs doesn't exist.