Add functionality for changing ES backend and implement Humio backend
Create an interface for es functionality (ESBackend) and
implement it for elasticsearch (ESClient) and humio (HumioClient).

The HumioClient implementation replicates functionality of es_*
functions in `rules/es_graphs.py` without changing the API.

ESClient is a wrapper for es_graphs.py functions.

Replace direct calls to es_* functions with calls to the chosen
implementation of the ESBackend interface (ESClient by default).

Minor fixes to the codebase; method calls on None objects and
other minor problems.

Co-authored-by: Potrik98 <potrik98@gmail.com>