feat(docker): add docker-compose.yml
We are adding a docker compose file to the project to make it easier to run the project locally. This docker compose run scirius with the appliance middleware by default. It uses a new "Dockerfile.local" to allow for local development in the container. We add a Dockerfile.appliance to build the appliance image. The repo "docker" folder was cleaned up to remove unused files.